pyproject.toml
[tool.poetry]
name = "travel-agent"
version = "0.1.0"
description = ""
authors = ["zjandali <yzjandali@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-aiplatform = ">=1.52"
vertexai = "^1.71.1"
python-dotenv = "^1.0.1"
logfire = "^3.1.0"
streamlit = "^1.41.1"
langchain = "^0.3.14"
pydantic-ai = "^0.0.18"
langchain-openai = "^0.3.0"
langgraph = "^0.2.62"
sendgrid = "^6.11.0"
serpapi = "^0.1.5"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
google-search-results = "^2.4.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
